make the boolean field label clickable again
This solution is both - ugly and more semantic. I moved the label around
the span elements, so now the displayed text is really a clickable
label. I also added an empty dummy label, so Uikit displays the pretty
switch again. It should be possible to avoid the uglyness by adding the
styles to `label:before` instead, but I don't want to touch Uikit right
now.