create bootstrap.php when creating folder via finder in addons dir
This issue came up so often and it is still a way to break the whole cockpit instance with a simple folder creation via ui.

Now the media api controller takes care of this and creates an empty `bootstrap.php` file after creating an empty folder inside the addons or the modules folder.

fixes #1215, fixes #1314